Added in some fine doctorate thesis on all aspects
This is an overview of all reasonable current ways of surveying land for construction or other various reasons. A good starting point for anyone wishing to build a mobile application for such projects.. If you wish to help us please message us we are Team Zero we always need more good hearted folks from all around the world.